+++ steps/export/utils.ts
@@ -10,13 +10,14 @@ import { hideBin } from 'yargs/helpers'
 import { fileURLToPath } from 'url'
 import { LanguageDescriptor, LanguageItemPair } from '../../lib/types.js'
 import { ContentProvider, Blob } from '@openzim/libzim/dist/index.js'
+import { formatLanguages } from 'steps/utils.js'
 
 dotenv.config()
 
 const __filename = fileURLToPath(import.meta.url)
 const __dirname = path.dirname(__filename)
 
-const argv: any = yargs(hideBin(process.argv)).array('includeLanguages').array('excludeLanguages').boolean('mulOnly').boolean('createMul').argv
+const argv: any = yargs(hideBin(process.argv)).string('includeLanguages').string('excludeLanguages').boolean('mulOnly').boolean('createMul').argv
 
 export const options = {
   catalogsDir: 'state/get/catalogs',
@@ -33,8 +34,8 @@ export const loadLanguages = async (): Promise<LanguageItemPair<LanguageDescript
   const langs = JSON.parse(langsFile.toString())
 
   return Object.entries(langs).reduce((acc, [key, value]) => {
-    if (argv.includeLanguages && !((argv.includeLanguages as string[]) || []).includes(key)) return acc
-    if (argv.excludeLanguages && ((argv.excludeLanguages as string[]) || []).includes(key)) return acc // noinspection RedundantIfStatementJS
+    if (argv.includeLanguages && !formatLanguages(argv.includeLanguages).includes(key)) return acc
+    if (argv.excludeLanguages && formatLanguages(argv.excludeLanguages).includes(key)) return acc // noinspection RedundantIfStatementJS
     return { ...acc, [key]: value }
   }, {})
 }