+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+* Support csv for --includeLanguages and --excludeLanguages and split values by ',' (#257)
+
 ## [3.1.0] - 2025-03-31
 
 ### Fixed